Next step for block-cyclic: set up domain classes
This commit takes the next step with the block-cyclic distribution
and sets up the state for the global and local domain classes.
To avoid a bunch of ugly modular arithmetic, I'm currently requiring
that the domain's low bound matches the one that was passed to the
distribution.  When debugging is on, each locale prints out the blocks
that it owns.  Updated the tests to add a domain to each test and print
it out.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16039 3a8e244f-b0f2-452b-bcba-4c88e055c3ca