Made --no-cpp-lines the default for -g with --devel/CHPL_DEVELOPER
Prior to this change, -g always turned on the --cpp-lines flag even
though developers typically prefer to debug in terms of the original C
code.  With this change, if --devel or CHPL_DEVELOPER is set, the
default will be --no-cpp-lines with or without -g.

Note that this dependence between flags is somewhat different than
others in that we normally process flags in the order that they are
specified, but in this case, we want being a developer to squelch the
cpp #line directives whether we set it first or last or through an
environment variable.  This resulted in somewhat more complicated
logic than I would have liked.  If you find a case that you think I
got wrong, let me know.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16042 3a8e244f-b0f2-452b-bcba-4c88e055c3ca