This change to the PVM comm layer changes how _chpl_message_info data is sent.
This change to the PVM comm layer changes how _chpl_message_info data is sent.
On a put or get, the data field in _chpl_message_info is the location of the
data. Before the change, it was sent as a stream of bytes. Now, it is sent
as an actual address using the CLASS_REFERENCE code.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16048 3a8e244f-b0f2-452b-bcba-4c88e055c3ca