This commit hopefully resolves hetero-PVM strings for good.
This commit changes the paradigm for sending and receiving strings in the
heterogeneous PVM case. A chpl_string in the course of a structure is
simply a pointer to a string (packaged often with the locale and the size).
However, when a string is sent directly via CHPL_COMM_WIDE_GET_STRING,
it is not a pointer. This latter case is represented internally via an
earlier submit with a negative number. Therefore, the change is: if we
want to send a string directly (negative in the table), do what has
always been done. However, if we're sending a pointer, use the
CLASS_REFERENCE functionality. This fix also removes the occasional
garbage characters on output.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16054 3a8e244f-b0f2-452b-bcba-4c88e055c3ca