This change adds a couple assertions to the PVM comm layer.
This change adds a couple assertions to the PVM comm layer. If we are
using ChplCommPut/Get/Fork on a negative size locally, something has gone
awry (we shouldn't be moving a string with this mechanism, for example).
If we get one of the cases I haven't implemented (which I've never seen
experimentally), we also raise an error.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16055 3a8e244f-b0f2-452b-bcba-4c88e055c3ca