Simplified logic to set CHPL_HOME and PVM_ROOT
It turns out I was overengineering my Makefile trick for
supporting CHPL_HOME and PVM_ROOT as strings in the
source code.  Furthermore, it seems the trick I was using
was fragile, as it resulted in the string "linux" getting
converted to "1" if it appeared in the CHPL_HOME path...
though I'm not sure what the source of the conversion was.
In any case this is simpler.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16057 3a8e244f-b0f2-452b-bcba-4c88e055c3ca