This fix allows the PVM launcher to run from archs with short names.
This fix allows the PVM launcher to run from architectures where the name,
say darwin, is longer than the one it is running to, say sunos-old. In
the algorithm I use to launch, in order to allow compiled files (for
different architectures) in similar directories, I replace every occurance
of the architecture name with the new one. In other words, /tmp/foo_linux64
will be replaced to /tmp/foo_linux in the appropriate case. However, the
replace_str algorithm I was using updated the buffer on the fly, and it
would overwrite the remainder of the buffer with extra characters from the
replacement string. This fixes this. It is possible we were seeing this
before from a linux node to a darwin node where there is one extra character
in darwin.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16067 3a8e244f-b0f2-452b-bcba-4c88e055c3ca