Changing Makefiles to use PrgEnv-gnu for GASNet when PrgEnv-cray is loaded
GASNet is known not to be compatible with the Cray compiler.  For
this reason, it was suggested (by David and/or Brian) that we build
GASNet with the gnu compiler when the target compiler is cray-xt-cray
and link (the object files are interchangeable).  This commit makes
this change by adding a new rule ENSURE_GASNET_COMPATIBLE_COMPILER
which is typically unset, but which for the cray-xt-cray compiler is
set to do a module swap.  We then use this command to configure and
build the GASNet directory and to build runtime/comm-gasnet/chpl_cfg_comm.c.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16080 3a8e244f-b0f2-452b-bcba-4c88e055c3ca