Fixed a termination bug in which termination could start even before all local destructors are called because calls to local destructors could be inserted after the call to downEndCount().  This fixes this issue in non-blocking forks in coforall statements.  There is still an issue (that is much rarer) in which the runtime could try to cancel a pthread that is not yet cancellable.
Fixed this bug for non-blocking forks by nesting the code that was
just before the call to downEndCount in its own block.  That way, the
destructor calls will be inserted before the downEndCount call.

This change should make

  multilocale/deitz/linked/test_linked3c.chpl

more stable.  It was consistently failing when run on two locales and
non-deterministically above that (the test setting is three).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16081 3a8e244f-b0f2-452b-bcba-4c88e055c3ca