Changed distributions from classes to values.
At the Chapel level, this means that all distributions must be created
by making a call to a new function called

  distributionValue(distributionClass)

that takes a newly allocated distribution class as an argument.  All
distribution types (which have not been specified to-date) must be
created by making a call to a new type function called

  distributionType(distributionClassType)

that takes a distribution class type as an argument.

Distribution Changes:

Added clone() methods to all distributions.  This required method
should return a new distribution class with the same value as 'this'.

Added getBaseDist() methods to all distributions that returns the
distribution class reference.

Implemented privatization for the Block distribution so that it is
more similar to Block domains and arrays.  Add a pid field and
replicate via a privatize() method.

Module Changes:

Like arrays and domains, distributions are now implemented with a
generic record, called _distribution, that wraps the underlying
classes.

Added autoCopy and autoDestroy calls for distributions that use a
reference count much like with arrays and domains.

Distributions are now destroyed much like domains were.  Domains
declared over distributions increment their reference count.  To
optimize the default distribution, disabled reference counting of
instances of DefaultDistribution by a parameter function
linksDistribution() that returns false on all domains distributed by
the DefaultDistribution.

Compiler Changes:

Added a flag "distribution" to mark this record type.

Fixed an issue with the code that (conservatively) frees values
allocated on the heap so that it no longer assumes a call is in the
IR (has a parentSymbol).

Specialized compiler code for argument passing and ".locale" so that
distributions behave like arrays and domains.

Test Changes:

Updated all tests to use calls to distributionValue as necessary.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16083 3a8e244f-b0f2-452b-bcba-4c88e055c3ca