This mod refactors the way config vars are handled during code generation, such that handling of internal (built-in) config vars is a little cleaner in the runtime.
If a config var is part of an internal modules, initialize it to be
part of module "Built-in" (this is known at compile time and can be
done during code generation).  This eliminates the need for any
runtime checks based on the name of the module being prepended with a
special string.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16086 3a8e244f-b0f2-452b-bcba-4c88e055c3ca