Change the name of a variable in the Block Array follower iterator.
The variable named "myLocArr" shadowed the privatized class member variable of
the same name making the optimization that checks if the array access is local
and avoids communication when it is local stop working. I changed the variable
name to "arrSection". The part of the array it describes is not necessarily
on the current locale, so this seems more appropriate than "myLocArr".

The optimization was commented out because of this problem, I uncommented
it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16101 3a8e244f-b0f2-452b-bcba-4c88e055c3ca