Update Cyclic distribution.
Renamed CyclicDist to Cyclic and LocCyclicDist to LocCyclic to follow the
pattern of Block and BlockCyclic. Used the filename as the module name instead
of declaring the module explicitly.

Changed the iterators to work with tuples of ranges so they can zip with
iterators from other distributions.

Added a "lowIdx" tuple to the distribution class that specifies an index that
will always be owned by the 0th locale in the targetLocs array. Previously,
the 0th locale always owned min(idxType).

Updated privatize functions to be more complete and added one for privatizing
the distribution class.

Removed the unused "alias" parameter.

Updated testCyclicDist for name changes and to give it a lowIdx of min(idxType)
so its behavior doesn't change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16110 3a8e244f-b0f2-452b-bcba-4c88e055c3ca