Updating Xcode project for Mac on behalf of Sam
Sam's log message:

Updated the Xcode project file: deleted files removed since last
update, added files new since last update, and deleted empty groups.
Tested only on a PowerMac G5, both by building from within Xcode, and
by building using util/makechpl, which is a deleted file.  (When built
with makechpl, the resulting compiler is not usable on a 64-bit
capable PowerPC-based Mac, because it only builds the runtime library
for the current architecture, which Xcode thinks is 32-bit PowerPC,
but the Chapel compiler explicitly creates 64-bit binaries when
compiling on a 64-bit capable PowerPC-based Mac.  On Intel-based Macs,
this is not a problem, because the Chapel compiler only ever creates
32-bit binaries, since there does not seem to be an easy way to
determine if the compiler is running on a 64-bit capable Intel-based
Mac.)

I also updated the README.macosx for the release to improve
expectations that the Xcode project would work but to warn that we
weren't particularly good at maintaining it between releases.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16111 3a8e244f-b0f2-452b-bcba-4c88e055c3ca