This commit adds another wrapper script called chpl-multirealm.
This commit adds another wrapper script called chpl-multirealm. The script
acts like the compiler in the sense that it simply passes all the same
arguments you'd use to compile a program to the compiler itself with some
pre-processing: find the realms of interest, append the name of the
architecture to the output file (and savec, if there is one), and compile
as normal.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16114 3a8e244f-b0f2-452b-bcba-4c88e055c3ca