Adding compiler/etc directory to release
Without this, users of the release can't use the --gdb
flag on generated binaries.  (Not that it's expected
that most will, but there's no reason to prevent them
from doing so).  This also adds the html helpers for
our --html view and the valgrind suppressions for the
compiler, both of which seem potentially useful to a
sufficiently motivated user.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16119 3a8e244f-b0f2-452b-bcba-4c88e055c3ca