This change reacts to the barrier before main in PVM comm code.
This change reacts to the barrier before main in PVM comm code. The barrier
code is designed to allow for safe exit; when src/main.c tells every non-0
node to exit, each of these nodes waits on a barrier. In PVM, this barrier
is blocking and non-thread-safe. To account for this, I set a flag in
rollcall to let nodes know if it is safe to barrier (essentially preventing
them from barriering until node 0 is ready). However, if an additional
barrier is added (as was the case the other day) between the rollcall and
the exit, then a hang could occur. Before this change, all programs hung
due to this new barrier. This change, while admittedly hacky, checks to
make sure the message matches the new barrier, and if so, everyone waits
correctly without hanging.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16125 3a8e244f-b0f2-452b-bcba-4c88e055c3ca