Changes to GASNET_MAX_SEGSIZE
The main change here is to change the units used to set
GASNET_MAX_SEGSIZE from unlabeled (which we think implies bytes) to
KB.  During the v0.9 release push, I wasn't able to get the kilobytes
version working (may have been doing something incorrect) and so
convinced myself that it wasn't necessary.  Later conversations with
the GASNet team pointed out that this would cause it to default to
Bytes and impact performance (surprisingly, it did not impact
performance which might suggest our memory allocator doesn't do well
with out-of-memory errors?  I didn't look into this question for this
commit).

The other change here is that since the v0.9 release, we created a new
chplcomm_md.h file for general GASNet communications that shadowed the
one used for the XT and caused GASNET_MAX_SEGSIZE not to get set.  I
worked around this for the time being by supporting overridable GASNet
files for both platforms (chplcomm_plat_md.h) and communication layers
(chplcomm_comm_md.h).

I need to test this on some other platforms -- sending patches from
Jaguar is not trivial.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16129 3a8e244f-b0f2-452b-bcba-4c88e055c3ca