Fixed a race in which the array of private objects that one thread is accessing could be bogus if another thread has reallocated it (by creating a new private object).  This introduces a leak of the array of private objects (but it is currently overwhelmed by the leak involving the private objects themselves). To fix this, we really need a data structure that we can lock when we write to it, but that we don't lock when we read from it.
Removed future for test

  parallel/sync/diten/userLevelEndCount

by removing the destructor of the user-level "EndCount" class.  It explicitly
deleted the sync variables in the class, but these should not be explicitly
deleted (at least anymore).

Removed future for test

  parallel/coforall/bradc/nested/nestedForall2

by adding error checking against using parallel constructs in serial iterators.
Added another test that uses a forall loop in a serial iterator.  This future
used a coforall in a serial iterator.

Fixed an issue with the data class type by replacing more accesses of the
substitutions vector with the getDataClassType function (see r15690).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16137 3a8e244f-b0f2-452b-bcba-4c88e055c3ca