Making sure we don't emit any classes or records with no fields
This is a codegen-time change to make sure we don't emit any
classes with no records or fields by inserting an "int" field
in the type's definition if it doesn't have any fields.  This
is unfortunate, but necessary because the Cray XT compiler
generates an error that seemingly cannot be squelched in such
cases.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16144 3a8e244f-b0f2-452b-bcba-4c88e055c3ca