Some Makefile fixes to better support cray-xt-cray
When compiling with cray-xt-cray, we need to be careful to only use
GASNet's C compile and link flags when we've swapped in the PrgEnv-gnu
compiler since the flags it decides to use may not be compatible with
Cray's compiler.  In the case of chpl_cfg_comm.c, this was easy to do
by simply special-casing the RUNTIME_CFLAGS variable in Makefile.share.
In compiling the generated code, the change is a bit bigger since we
normally add all of GASNet's flags into our C compile line.  Here, I
just took them out if the compiler was found to be cray-xt-cray.
Arguably, perhaps we shouldn't worry about GASNet's flags at this stage
in the compilation for any platform/compiler, but this was a bigger
change than I was willing to take on this close to the release.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16145 3a8e244f-b0f2-452b-bcba-4c88e055c3ca