Removed future for passing test
      test/classes/diten/record_initialize

Updated good as an error message since coforall is not allowed in a serial
iterator and removed future for test that should now pass

      test/functions/iterators/diten/zipCoforallIter

Updated execopt, updated expected leak (4 for extra element when array outside 
of function is resized), and serialized code to avoid leaking thread data for
the duration of the tracked function.  Removed future.

      test/memory/figueroa/LeakedMemory5.chpl

Recategorized bug future as an error message future.  I think iterating over a
loop-expression should be a non-var iterator and thus yield constant indices.
We don't check constness of indices in that the programmer can still assign an
index of a non-var iterator.  The effect of that assignment is just for the end
of the loop.  But perhaps this should be an error because the index is const.

      test/trivial/deitz/test_nested_var_for_expr

Recategorized bug future as semantic future.  Should a default constructor with
0-arguments be the default constructor when the type is specified.

      test/classes/figueroa/RecordConstructor1

Recategorized bug future to use a common entry "odd type name to partial
instantiation"

      test/classes/diten/instantiatedClassname

Word-smithed and clarified future categories for the following tests

      test/functions/deitz/iterators/test_no_yield
      test/functions/iterators/bradc/leadFollow/diffIndices

Recategorized string leaks futures to use a common entry "memory: string leaks".

      test/memory/figueroa/StringLeaks1
      test/memory/figueroa/StringLeaks3
      test/memory/figueroa/StringLeaks5
      test/memory/figueroa/StringLeaks2
      test/memory/figueroa/StringLeaks4

Recategorized more futures that will be fixed with out-of-order resolution.

      test/modules/diten/mutualuse2
      test/modules/diten/mutualuse3
      test/modules/diten/mutualuse


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16150 3a8e244f-b0f2-452b-bcba-4c88e055c3ca