Hard-coding XTs to use cc for GASNet compiles of generated code
This is necessary because when we build Chapel modules for the XT,
GASNet captures the location of the cc compiler used to build the
module, but when the module is installed, the cc compilers are
likely to be in a completely different place, breaking the compile
of the generated code.  In the past, we've worked around this by
having the chapel module set CHPL_MAKE to
"make GASNET_CC=cc GASNET_LD=cc ...", but this had the downside
that a user couldn't compile the generated code using a simple
"make -f output/Makefile" command.

This change also has a refactoring to my previous change which
turned of GASNet flags for the cray-xt-cray compiler to avoid
a doubly nested conditional.

David points out that he isn't really comfortable with using the
GASNet compiler with hard-coded path for non-XT compiles either,
but since the GASNet team has advocated doing that and since it's
a bigger change (conceptually -- it's easy to implement), I haven't
taken that on in this commit.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16224 3a8e244f-b0f2-452b-bcba-4c88e055c3ca