Finished up some work in ABQ that I mostly finished in IAD (all but the _local parameter so that this distribution can be used with COMM=none).
Added a new distribution called Private.  This is implemented with a single
distribution, domain, and array class.  There are no 'local' classes.  There is
a single data element per locale.  Added a new domain called PrivateSpace that
is a 1D arithmetic domain equivalent to LocaleSpace except that each index maps
to a unique locale.

At this point, a user of this distribution only need to deal with PrivateSpace.
It may be useful to create other distribution instances over different sets of
locales, but the distribution is currently hardcoded over the Locales array.  I
think there are design choices that we will need to make as to how to deal with
multiple realms and the implementation of the Locales arrays.

We can now create the Locales array over PrivateSpace, and even rename
PrivateSpace to LocaleSpace, though there is a question as to how to bootstrap
this since PrivateSpace is created using the Locales array.  We can either use
the primitive on-statements to create PrivateSpace or use an internal Locales
array.  Also if we change the Locales array to be over PrivateSpace, it would
be possible to turn locales into values instead of objects.

Added a parameter method requiresPrivatization() to distribution, domain, and
array classes that is similar to supportsPrivatization() except privatization
on these distributions, domains, and arrays is not disabled if
--no-privatization is used.  This method is used for the new PrivateDist
distribution.  Privatization is only disabled if --local is used.

Added an internal global parameter to the runtime called _local that is true if
we are compiling with --local, and false otherwise.  This is similar to
_privatization which is true if we are compiling with --privatization and
--no-local, and false otherwise.  I needed this to guard the use of a primitive
that only applies in the --no-local case.

Added a test that makes use of an array declared over PrivateSpace to compute
the local sums of a Block-distributed array.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16226 3a8e244f-b0f2-452b-bcba-4c88e055c3ca