Addressing some const char* issues in strstr, strchr, strrchr
User Sylvain Henry pointed out that he got warnings on Linux 2.6.31
and gcc 4.4.1 due to [const] char* of arguments and return values
in str*() routines.  C++ now supports two varieties of each call,
one that takes and returns const char*'s and one that takes and
returns char*s.  I adopted his patch for parser.cpp and files.cpp,
but took a different approach for driver.cpp to try and be a bit
more accurate (I think previously we were using a char* pointer to
modify a const char* buffer which seems inappropriate).  Sylvain
pointed us to a discussion on this issue at:

        http://gcc.gnu.org/ml/gcc-help/2009-08/msg00254.html

See the chapel-bugs archives for the thread on this topic.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16231 3a8e244f-b0f2-452b-bcba-4c88e055c3ca