Enabled CHPL_COMM_HETEROGENEOUS for CHPL_COMM == pvm correctly
In my commit the other day, I tried to enable CHPL_COMM_HETEROGENEOUS
for PVM, but only succeeded in doing it for files in the comm-pvm/
directory, which resulted in problems/deadlocks in multilocale
executions.  Here, I'm taking the setting up to the top-level
Makefiles so that they will be applied to all runtime sources.  This
makes hello-multiloc work for me again.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16234 3a8e244f-b0f2-452b-bcba-4c88e055c3ca