Optimized the expression ".locale.id" which before this change would evaluate .locale by determining the locale id of an expression and indexing into Locales (via Realms) to get the locale, and then evaluate .id by getting the id out of that locale.  With this change, this expression now just gets the locale id. Changed one line in the BlockDist code to compare locale IDs rather than locales to take advantage of this optimization.
Added a function to the Block distribution's domain class that returns a local
domain class to avoid returning the array of locale domain classes.

The net result of these two changes is that we no longer launch any remote
tasks back on to locale 0 during distribution, domain, and array setup.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16236 3a8e244f-b0f2-452b-bcba-4c88e055c3ca