Optimized the aligned Block-distributed array follower so that it takes a slice of its default arithmetic array field and iterates directly over that slice rather than iterating over the domain and indexing into this array.
Optimized the Default-distributed arithmetic array serial iterator if
the rank of the array is one so that it iterates over a computed range
and indexes into its data vector directly rather than iterating over
its domain and indexing into the array via the 'this' method.  This
eliminates much complexity due to the potential for random indexing.
This is a close approximation to the walker-bumper style optimization
done by the ZPL compiler.

Added the 'dont disable remote value forwarding' pragma to part of the
array slicing function so that we do not disable this important
optimization in the presence of array slicing even though the
reference counts are accessed.  This is important for keeping
iteration over block arrays fast because slicing is now done at the
beginning of the follower iterator.

Moved fragmentation of local blocks so that it happens after for loops
are expanded but before iterators are lowered.  It used to be before
for loops are expanded.  This makes it apply to the iterator of a for
or forall loop if the local block surrounds this for or forall loop.
This is important for the aligned Block-distributed array follower
because there is a local block around a loop.

The net effect of this change is a substantial performance improvement
of the stream benchmark on one locale with one task per locale.  On
jaguar, I was seeing 2.27 GB/s and I'm now seeing 3.74 GB/s.  This
brings us back to (or almost back to) the performance we were seeing
earlier this year, before the 'reindexed' parameter was permanently
set to true and performance over non-slices was optimized.  This is a
big step forward in that performance is optimized even over slices
now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16240 3a8e244f-b0f2-452b-bcba-4c88e055c3ca