Pair programmed with Alex Duran to implement Nanos implementation of tasking runtime
The trip to UPC was a success -- Alex and I were able to get the tasking layer
(equivalent to threads-generic, soon to be renamed tasks-generic with any
luck (or maybe even tasks-fifopool?)) over top of their next version of the
Nanos library to the extent that we could compile and run a multitasking
hello world program and the producer consumer example.  This implementation
has the benefits of being simple and of being able to put aside blocked tasks
in order to run other tasks -- so, for example, we can run fairly arbitrary
Chapel programs using a single thread -- something we have wanted to do for
quite some time!

We did cut a few corners -- we did not implement the serial state, nor most of the statistic functions,
 nor the task report or block report.  We also have not yet tested with GASNet,
which would be interesting. Also, some dead code cleanup should be done.

It still relies on the mutex implementation of the pthreads layer but this can be easily solved.

Now we need Alex to release a copy of Nanos that he is comfortable with so
we can check it into our third-party directory...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16252 3a8e244f-b0f2-452b-bcba-4c88e055c3ca