This change makes a few changes to the multirealm wrapper script.
This change makes a few changes to the multirealm wrapper script. 1) If you
set -o <some directory>, you get an error instead of a new file_<arch> named
the same thing as the directory. 2) Make sure the remote CHPL_HOST_PLATFORM
is set. 3) Make sure the remote CHPL_TARGET_PLATFORM is set. 4) The script
remotely cds to the $PWD, so if you don't include a path (and assume the
current working directory), it uses that.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16254 3a8e244f-b0f2-452b-bcba-4c88e055c3ca