This change fixes a problem with the PVM launcher in multirealm execution.
This change fixes a problem with the PVM launcher when the launcher is on
a realm where there are no locales. For example, if you were to define
your realmTypes to be linux and sunos-old, and you ran on 0 linux nodes from
a linux node (and at least one sunos-old node), the program would hang
infinitely (trying to launch a linux compiled app on the Sun node). The
change allows this possibility. As always, you must have at least 1 node
defined somewhere; you can't set your localesPerRealm to 0 and 0.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16256 3a8e244f-b0f2-452b-bcba-4c88e055c3ca