Removed the numRealms config const
At some point, Lee and I noticed that the numRealms config
param was fairly redundant with the realmTypes config param,
since they had to agree and the realmTypes value is
necessary in any case.  So this change defines numRealms
to be a param whose initial value is based on counting the
number of realms specified by the realmTypes config const.
Note that the realmTypes config param takes its default
value from CHPL_TARGET_PLATFORM, so this suggests that
you can also specify a multirealm compilation by setting
that environment variable to multiple platforms (e.g.,
setenv CHPL_TARGET_PLATFORM 'linux darwin sunos')

Removed several tests that were based on whether or not
numRealms matched realmTypes in number and such, which
are now bogus.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16266 3a8e244f-b0f2-452b-bcba-4c88e055c3ca