Support for multirealm stream and RA
This checkin commits versions of stream and RA that run
using multiple realms, using a helper module, AllLocales
which defines an array which includes all locales across
all realms.  We then change the distribution declarations
themselves to target the AllLocales arrays and the rest
of the code stays the same.  I should mention that I've
only tested this using GASNet multirealm runs, not PVM
and in particular, not heterogeneous architectures.

A few fixes were required in order to make this work.
The first was to expand the ChapelLocale.chpl int-to-locale
function to support multiple realms (when I first wrote
multirealm code I thought we didn't need this, but either
things have changed or I was wrong); the second is to
change the BlockDist and ChapelArray modules to do their
optimizations based on a locale's unique ID (uid) rather
than its normal ID (id) which matches its index in its
realm's LocaleSpace and therefore may not be unique
across locales.  The effect of using id was that some
things would be asserted local when they were not (resulting
in a runtime violation) or that nil pointers would be
dereferenced.

Next major steps here will be to see whether this AllLocales
array can and should be promoted to the internal modules,
and whether we ought to renumber each Realm's LocaleSpace
such that there's no need for a distinct id and uid per
locale.  Rather, they could be one and the same.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16275 3a8e244f-b0f2-452b-bcba-4c88e055c3ca