Added some new global variables to describe the complete locale set
This change adds three new global variables that describe the
complete set of locales available to a program across its
multiple realms:

const totNumLocales: int(32) -- the total number of locales
const AllLocaleSpace: domain(1) -- the space of all locales
const AllLocales: [AllLocaleSpace] locale -- the array of all locales,
                                          -- stored in sorted order,
                                          -- with Realm[0]'s locales
                                          -- first, then Realm[1]'s,
                                          -- and so forth

The goal of this change is to provide an easy way to talk about
all of the locales for cases where that is useful/helpful (such
as distributing an array over a heterogeneous set of workstations)
and also for use for some internal purposes (such as setting up
the tree of locales that Steve introduced earlier today).  It
also turns out to simplify program startup somewhat (by making
all locales get created in one spot) while making part of it
slightly more confusing (realms now get set up second, so have
to fill some information about themselves back into each locale).

In making this change, I also made each Realm's array of Locales
alias a slice of the AllLocales array.  Eventually the hope is
that we will privatize the AllLocales array so that each locale
has a private copy of all of the locales' global references.

Updated the HPCC tests I introduced earlier to refer to this array.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16281 3a8e244f-b0f2-452b-bcba-4c88e055c3ca