Took a pass over the multirealm test
Cleaned up comments and reorganized a bit to address new
features added yesterday (AllLocales array, e.g.), try
to show off certain concepts a bit more, etc.

Hardcoded the realmTypes to be linux64 in the compopts
file because this doesn't prevent it from working on
other platforms over GASNet (for nightly testing) and 
should match what we do our PVM testing on.  Ultimately, 
I'd like to add multirealm support to the testing system in
a more natural manner, as we do for multilocale.  Due
to the lack of this support, added a skip file for
CHPL_COMM==none testing for the time being.

Changed slightly how realms print themselves out to avoid
exposing the internal field name.  Updated the main
hello-multirealm test to reflect this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16287 3a8e244f-b0f2-452b-bcba-4c88e055c3ca