Simplifying PVM spawn logic
This change simplifies the PVM spawn logic significantly over the
previous version.  It also fixes a small problem that David found in
which single-realm executions that did not have CHPL_HOST_PLATFORM set
would set fault.

The new PVM spawn logic is as follows when launching on platform
"hostplatform" via:

  <path-to-foo>/foo

* If CHPL_MULTIREALM_LAUNCH_DIR_<platform> is set, the spawner will try
  to launch $CHPL_MULTIREALM_LAUNCH_DIR_<platform>/foo_real

* Otherwise, it will assume a symmetric file system and...

  * launch <path-to-foo>/foo_real if <path-to-foo> is absolute
  * $PWD/<path-to-foo>/foo_real if <path-to-foo> is relative

In all cases, occurrences of <hostplatform> within path-to-foo and foo
will be replaced by <platform>.

The other change simply always uses chpl_realmType() to determine the
type of the ream, whether in a single- or multi-realm execution.

Also removed a couple of dead malloc'd variables



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16312 3a8e244f-b0f2-452b-bcba-4c88e055c3ca