This change adds a table approach to chpl-multirealm for output.
This change adds a table approach to chpl-multirealm for output. First,
if the CHPL_MULTIREALM_LAUNCH_DIR_$platform env variable is set, it uses
that. If there are slashes in what was given to -o, then it gives and error
and continues (but doesn't compile for that platform). If not, then it uses
the name (or a.out) and works like before. If
CHPL_MULTIREALM_LAUNCH_DIR_$platform is not set, then it works like it always
did. Note, there's a poor hack to account for sunos-old in place that will
change shortly.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16321 3a8e244f-b0f2-452b-bcba-4c88e055c3ca