This change to the launcher allows start_test to work with hetero-PVM.
This change to the PVM launcher allows start_test to work correctly when the CHPL_MULTIREALM_LAUNCH_DIR_<arch> variables are set. It relies on CHPL_TEST_INVOKE_DIR, which is set by start_test. There's still the possibility a user would set this, but the worst case scenario is start_test would act as it did before this change.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16324 3a8e244f-b0f2-452b-bcba-4c88e055c3ca