Backing out the check for CHPL_TEST_INVOKE_DIR
From what I'm seeing in my testing, it's not necessary, and on
principle the runtime should not be aware of the testing system's
environment variables.  I believe that if the source of the problem
was CHPL_MULTIREALM_LAUNCH_DIR being set then it should either be
unset manually or in start_test.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16326 3a8e244f-b0f2-452b-bcba-4c88e055c3ca