This change in the PVM comm layer removes the termination lock.
This change in the PVM comm layer removes the termination lock, which is
unnecessary given that the chpl_comm_barrier (in conjunction with flags)
is responsible for successful termination. This commit also changes the
names of the lock wrappers which were originally named for their main use,
but are clearer now using the finishing lock state as the name.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16331 3a8e244f-b0f2-452b-bcba-4c88e055c3ca