This change makes two performance improvements to the hetero-PVM code.
This change makes two performance improvements to the hetero-PVM code. First,
it sets up a direct TCP port between tasks s.t. the pvmd is avoided (which
saw a significant but not overwhelming improvements in a couple atomic RA
tests). Second, it calls sched_yield in any chunk of code that's polling for
a message. This saw a fairly dramatic improvement in performance.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16340 3a8e244f-b0f2-452b-bcba-4c88e055c3ca