This commit add support for recursive leader iterators by changing such iterators into recursive functions that take loop body functions as arguments. This support is initial.  One issue I'm seeing is that a future I filed related to recursive leader iterators is not helped with this commit even if I add a base case to the recursive leader iterator in this test.
Details
-------

Runtime:

Added chpl_fn_args as a type alias of void* to the runtime.  It is used for
code generation of the generic function arguments to a loop body function.
This is linked to the compiler via dtFnArgs.

Compiler:

Added a transformation that changes recursive leader iterators into recursive
functions that take as arguments nested functions that capture the
functionality of loop bodies.  The nested functions are wrapped in wrapper
functions that take the outer function arguments in a bundle.  This extra level
of indirection appears necessary to allow similar call sites to share the same
code (and thus break an infinite loop).  Note also that we don't have to handle
loops over a recursive iterator from within a recursive iterator.  These only
need to be handled from within the transformed iterator function.  Refactored
iterator inlining code to share functionality.

Added a new class FnType that is a subtype of Type.  Such a class used to exist
as a future means of supporting first-class functions in Chapel.  Added a
function createFnType that creates a new FnType (with an associated TypeSymbol)
given a function symbol.  The only field in the FnType class is a FnSymbol.

Modified pruning of unused functions so that it accounts for functions that are
called only via a symbol of a function type (a function pointer).

Modified flatten functions so that it passes symbols in the outer function
without a reference if they don't have a reference (necessary for function
types and function argument bundles).

Modified dead code elimination so that indirect function calls (calls of
symbols of function types) are marked as essential.

Loosened actual-formal type mismatch checking to allow an actual of any
function type to be passed to a formal of any function type and to allow
anything to be passed to dtFnArgs.

Added codegen method on TypeSymbol that prints out the type name or chpl_fn_p
for function types.

Added specialized code generation for call expressions of arguments or
variables that are function types (function calls via function pointers).
Added a cast at code generation that casts an actual of a function type to a
generic function pointer.

Added extra verification to VarSymbol to make sure that VarSymbol::type is not
NULL.  This helped me debug during this change.

Added code generation of a comment for primitives that should no longer be in
the AST.  This allows code generation in the presence of the compiler flag
--ignore-errors which is useful for debugging.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@16341 3a8e244f-b0f2-452b-bcba-4c88e055c3ca